// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        DocumentPage: {
            /** Pagenumber */
            pageNumber: number;
            /** Blocks */
            blocks: components["schemas"]["PageBlock"][];
        };
// ...
        NormalizedBBox: {
            /** X */
            x: number;
            /** Y */
            y: number;
            /** Width */
            width: number;
            /** Height */
            height: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}